Move atom tests to integration directory
